css-in-jss: rename community-connect-ui folder
