+++ src/components/AdminPage/AdminPageLayout.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { breakPoints, spacing } from "@community-connect-ui/constants";
+import { breakPoints, spacing } from "community-connect-ui/constants";
 
 export const AdminPageWrapper = styled('div')`
     display: grid;
